Surface inputs and outputs inside padding in layouts
The existing solution for padding had several problems.
It could force an underground belt to need to stay underground longer than it actually had to,
since it could add an extra unusable "bus line" to the bus width.

Also, it made for an annoying interface for processes which must always bring the line
up from underground on their left-most column. Failure to do this might sometimes still work
but then randomly fail due to bus arrangements.

Our solution is to move surfacing into the hands of the layouter itself. To do this,
padding changed from 0-2 wide to 1-3 (with 0 becoming 3). We take care of raising
all inputs and outputs from underground in the left-most column, then pass them along
through the remaining padding.