Refactor how special cases like rockets and inf science are handled
The existing solution had two problems:
* Not accurate once modules taken into account
* Generator thinks that it needs to have a "rocket" step that outputs a rocket on a line,
then feed it into a second step that takes a rocket and a satellite.

We add two new features:
* An "is virtual" flag, which says that an output isn't actually made by this process,
  and it doesn't need to be moved on a bus. For now, we'll require virtual processes can't be inputs.
  We might want to relax that later if we want to account for power as an input.
* A "delay" and "fixed inputs", which aren't affected by modules.
We can then say that launches have a 41 second delay, reducing throughput to compensate for
the 41 second cutscene where a rocket is launched.
We can also include a satellite as an input to 1000 white science, without mistakenly thinking
it's affected by productivity like the rest of the inputs (the rocket parts) are.