Improve roboport row layout
The first roboport was half the distance out that it could be, as we were failing to take
into account the infra column roboport's range. This actually removes a lot of the special
casing and simplifies things.

Also, we now allow zero roboports if the width is really small enough.
We should check for this condition and just eliminate roboport rows entirely
for small factories, but this can be done later as it is a bit involved.